Remove linebreaks from beginnings of messages
Linebreaks are hard to see and translate on translatewiki.

As far as I can see, the messages with the linebreaks in the beginning
are used in two other messages, and it's easier to insert the linebreaks
into them.

To ensure that they are concatenated correctly, make implode
separate them with two linebreaks.