Fix file verification in upload form
File form components can't be required (explicitly in the form) due to
some upstream weirdness.
Instead we'll have to check to see if the file is empty when processing
the form in onSubmit.
UploadBase::getFile() is a no-op (it doesn't do anything, so we can
replace it with our isEmptyFile check.